requirements.txt
matplotlib
Pillow
requests